+++ tests/test_nasa_connect.py
@@ -1,4 +1,3 @@
-import pytest
 from src.nasa_connect import NasaConnect
 from src.twitter_connect import TwitterConnect
 
@@ -23,17 +22,3 @@ def test_can_obtain_title_and_url():
     assert nasa_connect_data["title"] == "Partial Solar Eclipse over Argentina"
     assert nasa_connect_data["url"] == "https://apod.nasa.gov/apod/image/2205/PartialEclipse_Andrada_960.jpg"
 
-def test_can_delete_tweets():
-    test_tweet = "Gonna Delete This"
-    twitter_connect = TwitterConnect()
-    response_data = twitter_connect.post_text(test_tweet)
-    delete_response = twitter_connect.delete_tweet(response_data["id"])
-    print(delete_response)
-    assert delete_response._json["id_str"] == str(response_data["id"])
-
-def test_can_post_to_twitter():
-    test_tweet = "Gonna Post Something"
-    twitter_connect = TwitterConnect()
-    response_data = twitter_connect.post_text(test_tweet)
-    assert response_data["text"] == test_tweet
-    twitter_connect.delete_tweet(response_data["id"])
